quick-reply changes
